Add the ability to use the 'rateOf' libsbml construct.
We have access to all the rates of change that we need in our models; this adds the ability to access them in the math of the model itself.

It doesn't quite work yet, but the basics are now there.  Checking this in so I can work on it later.